refactor: separated interfaces from domain into repository.ts and service.ts
