+++ src/presentation/controller/CreateUserController.ts
@@ -1,13 +1,13 @@
 import { Request, Response } from 'express'
-import { IServiceCreateUser } from 'interfaces/domain'
+import { ICreateUserService } from 'interfaces/domain/service'
 import { ICreateUserController } from 'interfaces/presentation/controller'
 import { inject, injectable } from 'tsyringe'
 import AbstractController from './AbstractController'
 
 @injectable()
 export default class CreateUserController extends AbstractController implements ICreateUserController {
-  service: IServiceCreateUser
-  constructor (@inject('CreateUserService') service: IServiceCreateUser) {
+  service: ICreateUserService
+  constructor (@inject('CreateUserService') service: ICreateUserService) {
     super(service)
     this.service = service
   }
